database-setup-main/requirements.txt
boto3==1.28.26
botocore==1.31.26
distlib==0.3.7
filelock==3.12.2
jmespath==1.0.1
platformdirs==3.9.1
PyMySQL==1.1.0
python-dateutil==2.8.2
python-dotenv==1.0.0
s3transfer==0.6.1
urllib3==1.26.16
virtualenv==20.24.1
